Added some classes, basic foot pressure functionality
