+++ lib/angles.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class Angles extends StatelessWidget {
+  int shk;
+  int hka;
+
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    return null;
+  }
+}
