+++ lib/rep.dart
@@ -1,0 +1,9 @@
+import 'package:nucoach/angles.dart';
+import 'package:nucoach/foot_pressure_map.dart';
+
+class Rep {
+  int repid;
+  int score;
+  Angles angles;
+  FootPressureMap fpmap;
+}
