+++ lib/session.dart
@@ -1,0 +1,4 @@
+class Session {
+  DateTime date;
+  List<Set> sets;
+}
