+++ lib/set.dart
@@ -1,0 +1,9 @@
+import 'package:nucoach/exercise.dart';
+import 'package:nucoach/rep.dart';
+
+class Set {
+  int setid;
+  Exercise exercise;
+  int score;
+  List<Rep> reps;
+}
